# src/config/openapi.yml
openapi: 3.0.3
info:
  title: |
    The Jolokia Api Server
    (Generated by `yarn build-api-doc`.)
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 0.1.2
  description: |
    This document contains a list of currently available apis that
    can be used by the **activemq-artemis-self-provisioning plugin**
    to get access to the [management api](https://github.com/apache/activemq-artemis/blob/main/docs/user-manual/management.adoc#the-management-api) of a broker instance via 
    its [jolokia](https://jolokia.org/) endpoint.

    ### How to add new apis to the src

    The api server uses [openapi](https://www.openapis.org/) to define
    its apis. All the apis are defined in the file [openapi.yml](src/config/openapi.yml).

    To add a new api first open the [openapi.yml](src/config/openapi.yml)
    and add a new api definition, for example:

    ```yaml
    /hello:
      get:
        description: hello api
        tags:
          - greeting
        operationId: hello
        parameters:
        - name: name
          required: false
          in: query
          description: The name of a caller
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DummyResponse'
      ```

    With the above section added, you defined the new api under the path `/hello`
...

...

...

...

...
    }
...
servers:
...
security:
  - bearerAuth: []
paths:
  /brokers:
    get:
      summary: retrieve the broker mbean
      description: >
        **Get the broker mbean**
          The return value is a one-element array that contains
          the broker's mbean object name.
      tags:
        - jolokia
      operationId: getBrokers
      parameters:
        - in: header
          name: endpoint-name
          schema:
            type: string
          required: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Broker'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /brokerDetails:
    get:
      summary: broker details
      description: >
        **Get the broker details**
          The return value is a json object that contains
          description of all the operations and attributes of the broker's mbean.
          It is defined in [ActiveMQServerControl.java](https://github.com/apache/activemq-artemis/blob/2.33.0/artemis-core-client/src/main/java/org/apache/activemq/artemis/api/core/management/ActiveMQServerControl.java)
      tags:
        - jolokia
      operationId: getBrokerDetails
      parameters:
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentDetails'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /readBrokerAttributes:
    get:
      summary: read broker attributes
      description: >
        **Read values of broker attributes**
          The return value is a json array that contains
          values of requested attributes of the broker's mbean.

          **Note**: to read multiple attributes, set it to **names** parameter
          separated by commas, e.g. `Version,Status`.
      tags:
        - jolokia
      operationId: readBrokerAttributes
      parameters:
        - name: names
          description: attribute names separated by commas. If not speified read all attributes.
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentAttribute'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /readAddressAttributes:
    get:
      summary: read address attributes
      description: >
        **Read values of address attributes**
          The return value is a json array that contains
          values of requested attributes of the addresses's mbean.

          **Note**: to read multiple attributes, set it to **attrs** parameter
          separated by commas, e.g. `RoutingTypes,Address`.
      tags:
        - jolokia
      operationId: readAddressAttributes
      parameters:
        - name: name
          description: the address name
          schema:
            type: string
          required: true
          in: query
        - name: attrs
          description: attribute names separated by commas. If not speified read all attributes.
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentAttribute'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /readQueueAttributes:
    get:
      summary: read queue attributes
      description: >
        **Read values of queue attributes**
          The return value is a json array that contains
          values of requested attributes of the queue's mbean.

          **Note**: to read multiple attributes, set it to **attrs** parameter
          separated by commas, e.g. `RoutingTypes,Address`.
      tags:
        - jolokia
      operationId: readQueueAttributes
      parameters:
        - name: name
          description: the queue name
          schema:
            type: string
          required: true
          in: query
        - name: address
          description: the address name
          schema:
            type: string
          required: true
          in: query
        - name: routing-type
          description: the routing type
          schema:
            type: string
          required: true
          in: query
        - name: attrs
          description: attribute names separated by commas. If not speified read all attributes.
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentAttribute'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /readAcceptorAttributes:
    get:
      summary: read acceptor attributes
      description: >
        **Read values of acceptor attributes**
          The return value is a json array that contains
          values of requested attributes of the acceptor's mbean.

          **Note**: to read multiple attributes, set it to **attrs** parameter
          separated by commas, e.g. `RoutingTypes,Address`.
      tags:
        - jolokia
      operationId: readAcceptorAttributes
      parameters:
        - name: name
          description: the queue name
          schema:
            type: string
          required: true
          in: query
        - name: attrs
          description: attribute names separated by commas. If not speified read all attributes.
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentAttribute'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /readClusterConnectionAttributes:
    get:
      summary: read cluster connection attributes
      description: >
        **Read values of cluster connection attributes**
          The return value is a json array that contains
          values of requested attributes of the cluster connection's mbean.

          **Note**: to read multiple attributes, set it to **attrs** parameter
          separated by commas, e.g. `NodeID, Topology`.
      tags:
        - jolokia
      operationId: readClusterConnectionAttributes
      parameters:
        - name: name
          description: the cluster connection name
          schema:
            type: string
          required: true
          in: query
        - name: attrs
          description: attribute names separated by commas. If not speified read all attributes.
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentAttribute'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /execClusterConnectionOperation:
    post:
      summary: execute a cluster connection operation
      description: >
        **Invoke an operation of the cluster connection mbean**

          It receives a POST request where the body
          should have the operation signature and its args.
          The return value is a one element json array that contains
          return values of invoked operation along with the request info.
      tags:
        - jolokia
      operationId: execClusterConnectionOperation
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: true
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRef'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecResult'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /execBrokerOperation:
    post:
      summary: execute a broker operation
      description: >
        **Invoke an operation of the broker mbean**

          It receives a POST request where the body
          should have the operation signature and its args.
          The return value is a one element json array that contains
          return values of invoked operation along with the request info.
      tags:
        - jolokia
      operationId: execBrokerOperation
      parameters:
        - in: query
          name: targetEndpoint
          schema:
            type: string
            pattern: '^https?:\/\/[a-zA-Z0-9@%._\+~#=-]{1,256}[:]?[0-9]*$'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRef'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecResult'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
...
components:
  schemas:
...
    Broker:
      type: object
      required:
        - name
      properties:
        name:
          type: string
...
    FailureResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [failed, error]
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer